+++ test_elasticsearch/test_server/test_helpers.py
@@ -2,10 +2,52 @@ from elasticsearch import helpers
 
 from . import ElasticTestCase
 
-class TestBulkIndex(ElasticTestCase):
+class TestStreamingBulk(ElasticTestCase):
     def test_all_documents_get_inserted(self):
         docs = [{"answer": x, '_id': x} for x in range(100)]
-        success, failed = helpers.bulk_index(self.client, docs, index='test-index', doc_type='answers', refresh=True)
+        for ok, item in helpers.streaming_bulk(self.client, docs, index='test-index', doc_type='answers', refresh=True):
+            self.assertTrue(ok)
+
+        self.assertEquals(100, self.client.count(index='test-index', doc_type='answers')['count'])
+        self.assertEquals({"answer": 42}, self.client.get(index='test-index', doc_type='answers', id=42)['_source'])
+
+    def test_all_errors_from_chunk_are_raised_on_failure(self):
+        self.client.indices.create("i",
+            {
+                "mappings": {"t": {"properties": {"a": {"type": "integer"}}}},
+                "settings": {"number_of_shards": 1, "number_of_replicas": 0}
+            })
+        self.client.cluster.health(wait_for_status="yellow")
+
+        try:
+            for ok, item in helpers.streaming_bulk(self.client, [{"a": "b"},
+                {"a": "c"}], index="i", doc_type="t", raise_on_error=True):
+                self.assertTrue(ok)
+        except helpers.BulkIndexError as e:
+            self.assertEquals(2, len(e.errors))
+        else:
+            assert False, "exception should have been raised"
+
+    def test_different_op_types(self):
+        self.client.index(index='i', doc_type='t', id=45, body={})
+        self.client.index(index='i', doc_type='t', id=42, body={})
+        docs = [
+            {'_index': 'i', '_type': 't', '_id': 47, 'f': 'v'},
+            {'_op_type': 'delete', '_index': 'i', '_type': 't', '_id': 45},
+            {'_op_type': 'update', '_index': 'i', '_type': 't', '_id': 42, 'doc': {'answer': 42}},
+        ]
+        for ok, item in helpers.streaming_bulk(self.client, docs):
+            self.assertTrue(ok)
+
+        self.assertFalse(self.client.exists(index='i', id=45))
+        self.assertEquals({'answer': 42}, self.client.get_source(index='i', id=42))
+        self.assertEquals({'f': 'v'}, self.client.get_source(index='i', id=47))
+
+
+class TestBulk(ElasticTestCase):
+    def test_all_documents_get_inserted(self):
+        docs = [{"answer": x, '_id': x} for x in range(100)]
+        success, failed = helpers.bulk(self.client, docs, index='test-index', doc_type='answers', refresh=True)
 
         self.assertEquals(100, success)
         self.assertFalse(failed)
@@ -14,7 +56,7 @@ class TestBulkIndex(ElasticTestCase):
 
     def test_stats_only_reports_numbers(self):
         docs = [{"answer": x} for x in range(100)]
-        success, failed = helpers.bulk_index(self.client, docs, index='test-index', doc_type='answers', refresh=True, stats_only=True)
+        success, failed = helpers.bulk(self.client, docs, index='test-index', doc_type='answers', refresh=True, stats_only=True)
 
         self.assertEquals(100, success)
         self.assertEquals(0, failed)
@@ -28,7 +70,7 @@ class TestBulkIndex(ElasticTestCase):
             })
         self.client.cluster.health(wait_for_status="yellow")
 
-        success, failed = helpers.bulk_index(
+        success, failed = helpers.bulk(
             self.client,
             [{"a": 42}, {"a": "c", '_id': 42}],
             index="i",
@@ -50,7 +92,7 @@ class TestBulkIndex(ElasticTestCase):
             })
         self.client.cluster.health(wait_for_status="yellow")
 
-        self.assertRaises(helpers.BulkIndexError, helpers.bulk_index,
+        self.assertRaises(helpers.BulkIndexError, helpers.bulk,
             self.client,
             [{"a": 42}, {"a": "c"}],
             index="i",
@@ -66,7 +108,7 @@ class TestBulkIndex(ElasticTestCase):
             })
         self.client.cluster.health(wait_for_status="yellow")
 
-        success, failed = helpers.bulk_index(
+        success, failed = helpers.bulk(
             self.client,
             [{"a": 42}, {"a": "c"}],
             index="i",